Complete module

+++ 17_todolist-v2/app.js
@@ -3,6 +3,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require('mongoose');
+const _ = require('lodash');
 
 const app = express();
 
@@ -58,7 +59,7 @@ app.get("/", function(req, res) {
 });
 
 app.get("/:customListName", function(req,res){
-  const customListName = req.params.customListName;
+  const customListName = _.capitalize(req.params.customListName);
 
   List.findOne({name: customListName}, (err, foundList) => {
     if (err) return;
@@ -105,12 +106,22 @@ app.post("/", function(req, res){
 app.post('/delete', (req, res) => {
   console.log(req.body);
   const checkedItemId = req.body.checkbox;
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (!err) {
-      console.log('Successfully deleted checked item.');
-      res.redirect('/');
-    }
-  });
+  const listName = req.body.listName;
+
+  if (listName === 'Today') {
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (!err) {
+        console.log('Successfully deleted checked item.');
+        res.redirect('/');
+      }
+    });
+  } else {
+    List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) => {
+      if (!err) {
+        res.redirect('/'+listName);
+      }
+    })
+  }
 });
 
 app.get("/about", function(req, res){
